Make backtracking stack allocate memory more efficiently
Replace simple linked-list stack with a chunk-based system. This makes the engine roughly twice as fast on my machine.

In this commit, each MatchingStackNode descendent is given a getSize() method. The next step in improving this will be to remove getSize() and replace it with having every popping method of every MatchingStackNode descendent pop itself, by passing its size to stack.pop().