Fix anchor optimization when doing more than one match on the same regex
On all matches subsequent to the first one (e.g. in stdin mode, lines after the first, or in numeric test mode, all test numbers after the first), the "anchored" boolean variable, which determines whether to apply an optimization that skips match attempts at positions other than 0, was uninitialized.

So, with some compilers and build configurations, this resulted in a bug in which regexes that were not anchored were evaluated as if they were anchored, when evaluating all matches after the first one.